fix: split CodeFactor Complex Method hotspots
Extract RuntimeEnv NDJSON/menu helpers and split Windows vs Unix Python
discovery so those methods stay under the complexity budget.
